start working to make the sidebar scroll again
